separate out docs sections to separate components
