+++ scopes/react/ui/compositions-summary/index.ts
@@ -1,0 +1,2 @@
+export { CompositionsSummary } from './compositions-summary';
+export type { CompositionsSummaryProps } from './compositions-summary';
