+++ scopes/react/ui/docs-content/index.ts
@@ -1,0 +1,2 @@
+export { DocsContent } from './docs-content';
+export type { DocsContentProps } from './docs-content';
