+++ scopes/react/ui/properties/index.ts
@@ -1,0 +1,1 @@
+export { Properties } from './properties';
